some type cleanup in pruning

+++ mockdown/pruning/conformance.py
@@ -3,8 +3,9 @@ from dataclasses import dataclass
 
 @dataclass(frozen=True)
 class Conformance:
-    width: int
-    height: int
+    # TODO: these should be ints! – Dylan
+    width: float
+    height: float
 
     # TODO: do we need x/y in top-level conformances? right now they're always 0
     x: int
